+++ src/auth/hooks/index.ts
@@ -1,3 +1,5 @@
 export { default as useAuth } from './useAuth';
 export { default as useInput } from './useInput';
 export { default as usePasswordIcon } from './usePasswordIcon';
+export { default as useSignIn } from './useSignIn';
+export { default as useSignUp } from './useSignUp';
